logger to file

+++ Application/Services/UsersService.cs
@@ -7,12 +7,16 @@ using AutoMapper;
 using Domain.Entities;
 using Domain.Entities.Users;
 using Domain.Interfaces;
+using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
+using System.Security.Claims;
 using System.Text;
 using System.Threading.Tasks;
+using System.Web.Http;
 
 namespace Application.Services
 {
@@ -21,53 +25,68 @@ namespace Application.Services
 
         private readonly IUsersRepository _usersRepository;
         private readonly IMapper _mapper;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public UsersService(IUsersRepository usersRepository, IMapper mapper)
+        public UsersService(IUsersRepository usersRepository, IMapper mapper, IHttpContextAccessor httpContextAccessor)
         {
             _usersRepository = usersRepository;
             _mapper = mapper;
+            _httpContextAccessor = httpContextAccessor;
         }
 
-        public UsersDto Authenticate(string username, string password)
-        {
-            var user = _usersRepository.Authenticate(username, password);
+       
 
-            return _mapper.Map<UsersDto>(user);
+        public bool CanCurrentUserManipulateData(Users existingUser, Users modifiedUser = null)
+        {
+            
+            ClaimsIdentity identity = _httpContextAccessor.HttpContext.User.Identity as ClaimsIdentity;
+            return CheckUserRoles.CanCurrentUserUpdate(identity, existingUser, modifiedUser);
         }
 
 
         public UsersDto GetById(int id)
         {
-            var user =_usersRepository.GetById(id);
+            var user = _usersRepository.GetById(id);
+            if (user == null || !CanCurrentUserManipulateData(user))
+               throw new BadRequestException(ResponseMessage.BadRequestForId);
+            return _mapper.Map<UsersDto>(user);
+        }
+        public UsersDto GetCurrentUser()
+        {
+            var user = _usersRepository.GetByUsername(_httpContextAccessor.HttpContext.User.Identity.Name);
             return _mapper.Map<UsersDto>(user);
         }
 
         public (IEnumerable<UsersDto>, PagedDto) GetAllWithFilters(ref PaginationDto paginationData, string username = null, RolesEnum? role = null)
         {
             Pagination pagination = _mapper.Map<Pagination>(paginationData);
-
             var users = _usersRepository.GetAllWithFilters(ref pagination, username, role);
-
             var paged = _mapper.Map<PagedDto>(pagination);
             return (_mapper.Map<IEnumerable<UsersDto>>(users), paged);
         }
 
         public UsersDto CreateUser(CreateUsersDto userData)
         {
-            var post = _mapper.Map<Users>(userData);
-            _usersRepository.CreateUser(post);
-            return _mapper.Map<UsersDto>(post);
+            var user = _mapper.Map<Users>(userData);
+            if (user == null || !CanCurrentUserManipulateData(user))
+                throw new BadRequestException(ResponseMessage.BadRequestForId);
+            _usersRepository.CreateUser(user);
+            return _mapper.Map<UsersDto>(user);
         }
 
         public void DeleteUser(int id)
         {
-            var user = _usersRepository.GetById(id);
-            _usersRepository.DeleteUser(user);
+                var user = _usersRepository.GetById(id);
+                if (user == null || !CanCurrentUserManipulateData(user))
+                    throw new BadRequestException(ResponseMessage.BadRequestForId);
+                _usersRepository.DeleteUser(user);
         }
 
         public void UpdateUser(int id, UpdateUsersDTO userData)
         {
             var existingUser = _usersRepository.GetById(id);
+            if (existingUser == null || !CanCurrentUserManipulateData(existingUser, _mapper.Map<Users>(userData)))
+                throw new BadRequestException(ResponseMessage.BadRequestForId);
             var user = _mapper.Map(userData, existingUser);
             _usersRepository.UpdateUser(user);
         }
